Initial Commit
First commit, improving control of saved code, etc. This mod is a refactor of the old, deprecated nc_bundy (old, now graveyarded, mod focused on introducing recolonization of rotting tree stumps exposed to still water). The symbiont concept still remains, though now less macabre. Perhaps one of the more interesting ideas, hopefully the code will catch up to the idea in terms of coolness. This mod was intended to be one of the principal 3 daughter mods for the Rhynia library (if it can even be called a library) (for the illustrious Nodecore (Minetest game)). It was the first attempt at implementing the adaptation the old Rhynia library to accept nodebox drawtype registered nodes rather than just plantlike drawtypes. Old functionality is much different now, as the code once made this mod independent was mostly changed for inclusion into Rhynia (what wasn't otherwise deleted). Eventually it's brother mod "nc_pollarding" will be migrated over as a module to Rhynia (plans are to add a registered plant-trait that allows plants to be coppiced and pollarded, rather than the old, crashy "pollarding resin" method).